chore: splits util file tests into files
